pyproject.toml
[project]
name = "mft-autocot"
version = "0.3.0"
description = "MFT-AutoCoT experiment with Hydra"
requires-python = ">=3.10"

dependencies = [
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "wandb>=0.16.0",
  "torch>=2.2.0",
  "transformers>=4.40.0",
  "datasets>=2.18.0",
  "sentence-transformers>=2.6.0",
  "scikit-learn>=1.4.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "optuna>=3.6.0",
  "scipy>=1.11.0",
  "pyyaml>=6.0.1"
]
